+++ app/src/main/java/com/lekstel/aqi/stations/data/model/StationOnMapDTO.kt
@@ -3,7 +3,7 @@ package com.lekstel.aqi.stations.data.model
 data class StationOnMapDTO(
     val lat: Double,
     val lon: Double,
-    val uid: Long,
+    val uid: Int,
     val aqi: String,
     val station: StationShortInfoDTO
 ) {
