+++ app/src/main/java/com/lekstel/aqi/stations/domain/model/StationOnMap.kt
@@ -3,7 +3,7 @@ package com.lekstel.aqi.stations.domain.model
 import com.lekstel.aqi.base.presentation.view.adapter.BaseListItem
 
 data class StationOnMap(
-    override val id: Long,
+    override val id: Int,
     val lat: Double,
     val lon: Double,
     val aqi: String,
